+++ libs/docs/platform/message-popover/examples/form-generator/form-generator-component-example.component.ts
@@ -61,7 +61,7 @@ export class FormGeneratorComponentExample {
                     placeholder: 'Please provide your name',
                     guiOptions: {
                         hint: {
-                            text: 'Some contextual hint',
+                            content: 'Some contextual hint',
                             glyph: 'accidental-leave'
                         },
                         appendColon: true,