+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# [0.40.0-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.40.0-rc.0...v0.40.0-rc.1) (2023-03-24)
+
+
+### Features
+
+* **platform:** form generator inline help support for templates ([#9560](https://github.com/SAP/fundamental-ngx/issues/9560)) ([5814ba3](https://github.com/SAP/fundamental-ngx/commit/5814ba3a2a2b920df7c5bb7b77bbbff9582c4fb2))
+
+
+
+# [0.40.0-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.39.1-rc.29...v0.40.0-rc.0) (2023-03-24)
+
+
+### Bug Fixes
+
+* **core,platform:** multi input and combobox tokenizer behaviour ([#9571](https://github.com/SAP/fundamental-ngx/issues/9571)) ([7430869](https://github.com/SAP/fundamental-ngx/commit/74308695ab2ae1b4c370431a6f612cc838119034))
+
+
+### BREAKING CHANGES
+
+* **core,platform:** Multi-combobox and Multi-input components for both libraries (core and platform) are now following [guidelines logic regarding reviewing tokens](https://experience.sap.com/fiori-design-web/multiinput/#reviewing-tokens)
+
+
+
 ## [0.39.1-rc.29](https://github.com/SAP/fundamental-ngx/compare/v0.39.1-rc.28...v0.39.1-rc.29) (2023-03-23)